requirements.txt
prompt_toolkit>=3.0.0
transformers>=4.35.0
torch>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0 